requirements.txt
google-genai
python-dotenv
rouge-score
polars
bert-score
ipywidgets
ollama
great_tables
openai
tikzplotlib
